+++ src/components/FindNotebooksCommand.tsx
@@ -13,7 +13,7 @@ import { bold, codeBlock, inlineCode, italic, quoteBlock } from "../markdown";
 
 import { copyShortcut } from "./shortcuts";
 import { ColorDefault, ColorEmphasis, ColorPrivate } from "./colors";
-import ExpandableErrorToast from "./ExpandableErrorToast";
+import ExpandableToast from "./ExpandableToast";
 
 const link = new LinkBuilder("notebooks");
 
@@ -33,7 +33,7 @@ export default function FindNotebooksCommand({ src }: { src: Sourcegraph }) {
 
   const { push } = useNavigation();
   if (error) {
-    ExpandableErrorToast(push, "Unexpected error", "Find notebooks failed", error.message).show();
+    ExpandableToast(push, "Unexpected error", "Find notebooks failed", error.message).show();
   }
 
   const srcName = instanceName(src);